% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afforestation.r
\name{afforestation}
\alias{afforestation}
\title{Simulation of afforestation in shrublands}
\usage{
afforestation(land, params)
}
\arguments{
\item{land}{Data frame with the description of the landscape: cell identifier, land-cover type, and
time since last cover-change}

\item{params}{List of the parameters of the model}

\item{coords}{Data frame with the identifier and coordinates of the raster cells}
}
\value{
Returns a list of two vectors with cells' identifiers afforested by oaks and pines, respectively, 
and a data frame with tracking data of the afforestation process
}
\description{
Simulation of afforestation in shrublands
}
\examples{
data(landscape)
data(coords)
params = default.params()
res = afforestation(landscape, coords, params)

}
