% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{burnt.area.dist.gz}
\alias{burnt.area.dist.gz}
\title{Annual area burnt in Galicia}
\format{
A data frame with as many rows as grid cells in the study area and 2 variables:
\describe{
  \item{upper.th}{Upper threshold of discrete intervals of annual area burnt}
  \item{freq}{Number of fires per class}
}
}
\usage{
burnt.area.dist.gz
}
\description{
Discrete distribution of annual area burnt in Galicia observed between 1986-2011
}
\keyword{datasets}
