% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{burnt.area.dist.pt}
\alias{burnt.area.dist.pt}
\title{Annual area burnt in Portugal}
\format{
A data frame with as many rows as grid cells in the study area and 2 variables:
\describe{
  \item{upper.th}{Upper threshold of discrete intervals of annual area burnt}
  \item{freq}{Number of fires per class}
}
}
\usage{
burnt.area.dist.pt
}
\description{
Discrete distribution of annual area burnt in Portugal observed between 1980-2016
}
\keyword{datasets}
