% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{coords}
\alias{coords}
\title{UTM coordinates}
\format{
A data frame with as many rows as grid cells in the study area and 3 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{x}{UTM coordinate X}
  \item{y}{UTM coordinate Y}
}
}
\usage{
coords
}
\description{
Coordinates x and y in UTM29N-ETRS89 of the 1 ha cells in the study area
}
\keyword{datasets}
