% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.params.r
\name{default.params}
\alias{default.params}
\title{Default values of the model's parameters}
\usage{
default.params()
}
\value{
Returns a list of the parameters of the model and the default values
}
\description{
Default values of the model's parameters
}
\examples{
params = default.params()

}
