% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encroachment.r
\name{encroachment}
\alias{encroachment}
\title{Simulation of encroachment in sparse vegetation}
\usage{
encroachment(land, params)
}
\arguments{
\item{land}{Data frame with the description of the landscape: cell identifier, land-cover type, and
time since last cover-change}

\item{params}{List of the parameters of the model}

\item{coords}{Data frame with the identifier and coordinates of the raster cells}
}
\value{
Returns a list of one vector with cells' identifiers colonized by shrublands and a data frame 
with tracking data of the encroachment process
}
\description{
Simulation of encroachment in sparse vegetation
}
\examples{
data(landscape)
data(coords)
params = default.params()
res = encroachment(landscape, coords, params)

}
