% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.risk.r
\name{fire.risk}
\alias{fire.risk}
\title{Compute fire risk as a function of landscape susceptibility, danger, and damage}
\usage{
fire.risk(land, orography, ha.cell)
}
\arguments{
\item{land}{Data frame with the description of the landscape: cell identifier, land-cover type, and
time since last cover-change}

\item{orography}{Data frame with the geographic data of the raster cells: cell identifier, region 
(\code{GZ} is Galicia and \code{PT} is Protugal), protected (\code{TRUE}/\code{FALSE}), elevation (in m), aspect
 (1 - north, 2 - east, 3 - south, 4 - west), dens.road (density of roads per km^2^), slope (in %), and utm 
 (code of the UTM 1x1 grid).}

\item{ha.cell}{Number of hectares each raster cell in the grid represents}
}
\value{
Returns a vector with the fire risk of the raster cells that is function fo the land-cover type and 
the orography
}
\description{
Compute fire risk as a function of landscape susceptibility, danger, and damage
}
\examples{
data(landscape)
data(orography)
res = fire.risk(landscape, orography, 1)

}
