% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fire.size.dist.pt}
\alias{fire.size.dist.pt}
\title{Fire size distribution in Portugal}
\format{
A data frame with two variables:
\describe{
  \item{upper.th}{Upper threshold of discrete intervals of fire size}
  \item{freq}{Number of fires per class}
}
}
\usage{
fire.size.dist.pt
}
\description{
Discrete distribution of annual area burnt in Portugal observed between 1980-2016
}
\keyword{datasets}
