% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.recover.r
\name{foerst.recover}
\alias{foerst.recover}
\alias{forest.recover}
\title{Simulation of encroachment in sparse vegetation}
\usage{
forest.recover(land, params, shrub2pine, shrub2oak, t = 1)
}
\arguments{
\item{land}{Data frame with the description of the landscape: cell identifier, land-cover type, and
time since last cover-change}

\item{params}{List of the parameters of the model}

\item{shrub2pine}{Vector to accumulate the number of shrub.to.pine cells that have to transform to pine each time step}

\item{shrub2oak}{Vector to accumulate the number of shrub.to.oak cells that have to transform to oak each time step}

\item{t}{Value of the current time step}

\item{coords}{Data frame with the identifier and coordinates of the raster cells}
}
\value{
Returns a list with (1) a vector of cells' identifiers that have transformed to pine, (2) a vector of cell's 
identifiers that have transformed to oak, (3) a vector with the number of cells to be transformed to pine each time step, 
(4) a vector with the number of cells to be transformed to oak each time step, and (5) a data frame with tracking data 
of the forest recover process
}
\description{
Simulation of encroachment in sparse vegetation
}
\examples{
data(landscape)
data(coords)
params = default.params()
shrub2pine = shrub2oak = numeric(length=params$time.horizon+8)
res = forest.recover(landscape, coords, params, shrub2pine, shrub2oak, 1)

}
