% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.cover.change.r
\name{land.cover.change}
\alias{land.cover.change}
\title{Simulation of land-cover changes}
\usage{
land.cover.change(
  land,
  params,
  lc.trans = 1,
  trgt.dmnd = 0,
  visit.cells = NULL
)
}
\arguments{
\item{land}{Data frame with the description of the landscape: cell identifier, land-cover type, and
time since last cover-change}

\item{params}{List of the parameters of the model}

\item{lc.trans}{Identifier of the land-cover transition: 1 - smart plantation, 2 - agriculture conversion, 
3 - rurla abandonment, and 4 - pasture abanonment.}

\item{trgt.dmnd}{Number of hecatres to be converted to the target land-cover type}

\item{visit.cells}{Vector of cells' identifier that have already been subject of change for a land-cover change process}

\item{orography}{Data frame with the geographic data of the raster cells: cell identifier, region 
(\code{GZ} is Galicia and \code{PT} is Protugal), protected (\code{TRUE}/\code{FALSE}), elevation (in m), aspect
 (1 - north, 2 - east, 3 - south, 4 - west), dens.road (density of roads per km^2^), slope (in %), and utm 
 (code of the UTM 1x1 grid).}

\item{coords}{Data frame with the identifier and coordinates of the raster cells}
}
\value{
Returns a vector with cells' identifiers to change the land-cover type
}
\description{
Simulation of land-cover changes
}
\examples{
data(landscape)
data(orography)
data(coords)
params = default.params()
res = land.cover.change(landscape, orography, coords, params, 1, 100, NULL)

}
