% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.dyn.mdl.r
\name{land.dyn.mdl}
\alias{land.dyn.mdl}
\title{REMAINS model}
\usage{
land.dyn.mdl(
  scenDir,
  is.land.cover.change = TRUE,
  is.wildfire = TRUE,
  is.prescribed.burn = TRUE,
  is.postfire.rege = TRUE,
  is.forest.recover = TRUE,
  is.afforestation = TRUE,
  is.encroachment = TRUE,
  nrun = 1,
  params = NULL,
  lcc.demand = NULL,
  save.land = FALSE,
  out.maps = FALSE,
  write.outputs = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{scenDir}{String with the name of the scenario that is the directory path to save the model's outputs}

\item{is.land.cover.change}{A flag to indicate that land cover changes are simulated}

\item{is.wildfire}{A flag to indicate that wildfires are simulated}

\item{is.prescribed.burn}{A flag to indicate that prescribed burns are simulated}

\item{is.postfire.rege}{A flag to indicate that post-fire regeneration is simulated}

\item{is.forest.recover}{A flag to indicate that forest recover after fire is simulated}

\item{is.afforestation}{A flag to indicate that afforestation is simulated}

\item{is.encroachment}{A flag to indicate that encroachment is simulated}

\item{nrun}{Number of replicates to run the model}

\item{params}{A list of the default model's parameters}

\item{lcc.demand}{A data frame with the area to be changed by each land-cover transition per time step}

\item{save.land}{A flag to save as a RDS file the \code{landscape} data frame at the time step indicated in \code{out.seq}}

\item{out.maps}{A flat to save maps as .tif files of wildfire identifier, wildfire step and fire risk}

\item{write.outputs}{A flag to write the results of the landscape processes as text files}

\item{verbose}{A flag to prompt extra information}
}
\value{
A list with the following 13 items:
 \itemize{
   \item{\code{Land}: A data frame of tree species abundance, volume under bark, volume with bark, 
   and carbon stock per age class, and area per non-forest land-cover types, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{spp}: Code of the tree species or land-cover type.}
        \item{\code{age.class}: Code of the age class: young, mature, or old.}
        \item{\code{area}: Area (in ha).}
        \item{\code{vol}: Volume under bark in \eqn{m^{3}} for tree species and biomass in tonnes for shrublands.}
        \item{\code{volbark}: Volume with bark in \eqn{m^{3}}.}
        \item{\code{carbon}: Carbon content (in Mg).}
      }
   }
   \item{\code{LandSQI}: A data frame of area and volumes per tree species and per SQI, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{spp}: Code of the tree species or land-cover type.}
        \item{\code{sqi}: Code of SQI: 1 - low, 2 - good, 3 - optimal.}
        \item{\code{area}: Area (in ha).}
        \item{\code{vol}: Volume under bark in \eqn{m^{3}}.}
        \item{\code{volbark}: Volume with bark in \eqn{m^{3}}.}
      }
   }
   \item{\code{Carbon}: A data frame of carbon emissions and generation per type of process
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{type}: Process involved in the generation or emission of Carbon dioxide.}
        \item{\code{carbon}: Carbon dioxide quantity in \eqn{g}.}
      }
   }
   \item{\code{Harvest}: A data frame of sawlog and wood volume harvested per species 
   (included if \code{is.harvesting}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{spp}: Code of the species.}
        \item{\code{vol.sawlog}: Harvested timber volume for sawlog in \eqn{m^{3}}.}
        \item{\code{vol.wood}: Harvested timber volume for wood in \eqn{m^{3}}.}         
      }
   }
   \item{\code{ForestArea}: A data frame of areas for sustainable timber harvesting
   (included if \code{is.harvesting}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{forest}: Area of all tree species.}
        \item{\code{spp.harvestable}: Area of harvestable tree species.}
        \item{\code{non.protect}: Harvestable area without any protection status.}
        \item{\code{national.park}: Harvestable area in a national park.}
        \item{\code{enpe}: Harvestable area protected, but not in a national park.}
        \item{\code{no.park}: Harvestable area not in a national park.}
        \item{\code{slope30.nopark}: Harvestable area not in a national park with slope <= 30 pct.}
        \item{\code{slope30.nopark.distpath1.5}: Harvestable area not in a national park with slope <= 30 pct and distance to roads or forest tracks <= 1.5 km.}
        \item{\code{slope30.nopark.distpath2.2}: Harvestable area not in a national park with slope <= 30 pct and distance to roads or forest tracks <= 2.2 km.}
      }
   }
   \item{\code{HarvestArea}: A data frame of area harvested per sylvicultural prescription and forest type    
   (included if \code{is.harvesting}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{todo}: Sylvicultural prescription: prep.cut, removal.cut, seed.cut, or thinning.}
        \item{\code{fytpe}: Forest type: conif or decid.}
        \item{\code{area}: Area in ha.}
      }
   }
   \item{\code{HarvestVolume}: A data frame of potential and total volum extracted per forest type
   (included if \code{is.harvesting}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{fytpe}: Forest type: conif or decid.}
        \item{\code{vol.potential.extract.sawlog}: Potential harvesting volume for sawlog in \eqn{m^{3}}.}
        \item{\code{vol.potential.extract.wood}: Potential harvesting volume for Wood in \eqn{m^{3}}.}     
        \item{\code{vol.extract.sawlog}: Harvested volume for sawlog in \eqn{m^{3}}.}
        \item{\code{vol.extract.wood}: Harvested volume for wood in \eqn{m^{3}}.}
        \item{\code{pct.sawlog}: Percentage of harvested volume for sawlog.}
        \item{\code{pct.wood}: Percentage of harvested volume for wood.}     
      }
   }
   \item{\code{Fires}: A data frame of target, burnt and suppresed area per fire
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{swc}: Synoptic weather condition: 1 - wind, 2 - heat, 3 - regular.}
        \item{\code{clim.sever}: Climatic severity: 0 - mild, 1 - severe, 2 - extreme.}
        \item{\code{fire.id}: Fire event identificator.}
        \item{\code{fst}: Fire spreading type: 1 - wind-driven, 2 - convective, 3 - topographic.}
        \item{\code{wind}: Main wind direction in degrees.}
        \item{\code{atarget}: Target area to be burnt (in ha).}
        \item{\code{aburnt.highintens}: Area burnt in high intensity (in ha).}
        \item{\code{aburnt.lowintens}: Area brunt in low intensity (in ha).}
        \item{\code{asupp.fuel}: Area suppressed in low-fuel conditions (in ha).}
        \item{\code{asupp.sprd}: Area suppressed in slow fire spread conditions (in ha).}
        \item{\code{rem}: Remanent area, not burnt, neither suppressed (in ha).}
      }
   }
   \item{\code{BurntSpp}: A data frame of burnt area and biomass per species or land-cover type
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{fire.id}: Fire event identificator.}
        \item{\code{spp}: Code of the tree species or land-cover type.}
        \item{\code{aburnt}: Area effectively burnt (in ha).}
        \item{\code{bburnt}: Basal area effectively burnt (in \eqn{m^{2}ha^{-1}}).}
      }
   }
   \item{\code{PostFire}: A data frame of species replacement after fire
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{spp.out}: Code of the tree species or land-cover type replaced.}
        \item{\code{spp.in}: Code of the tree species or land-cover type after replacement.}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{Drought}: A data frame of drought-induced mortality per tree species
   (included if \code{is.drought}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{spp}: Code of the tree species.}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{Cohort}: A data frame of species replacement after drought-induced mortality 
   (included if \code{is.cohort.establish}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{spp.out}: Code of the tree species replaced.}
        \item{\code{spp.in}: Code of the tree species or land-cover type after replacement.}
        \item{\code{area}: Area (in ha).}
      }
   }   
   \item{\code{Afforest}: A data frame of new tree species following shrubland colonization
   (included if \code{is.afforestation}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{spp}: Code of the tree species.}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{Encroach}: A data frame of new shrubland area following encroachment
   (included if \code{is.encroachment}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{spp}: Code of shrublands.}
        \item{\code{area}: Area (in ha).}
      }
    }
 }
}
\description{
Run the REMAINS model that includes the processes of land-cover changes, wildfires, prescribed burns, 
and vegetation dynamics in mountain landscapes.
}
\examples{

\dontrun{
library(medLDM)
# Run one single 90-year replicate with forest management
result = land.dyn.mdl(is.harvest = T)
}

}
