% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lcc.demand}
\alias{lcc.demand}
\title{Induced landscape changes}
\format{
A data frame with as many rows as grid cells in the study area and 3 variables:
\describe{
  \item{SmartPlant}{Conversion rate for smart plantations: oaks spp (in ha/year)}
  \item{AgriConver}{Conversion rate for agricultural conversion (in ha/year)}
  \item{RuralAbnd}{Conversion rate for rural abandonment (in ha/year)}
}
}
\usage{
lcc.demand
}
\description{
Induced landscape changes: rural abandonment, agricultural conversion and smart plantations
}
\keyword{datasets}
