% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mask.study.area}
\alias{mask.study.area}
\title{Mask of the study area}
\format{
Raster of 648 (nrow) x 791 (ncol)
\describe{
Raster of the study area (1 or NA) in the UTM29N - ETRS89 projection, at 1 ha of spatial resolution. 
The unique grid cell identificator \code{cell.id} in the \code{landscape} data frame coincides with 
the position of the location in the \code{mask} raster.
}
}
\usage{
mask.study.area
}
\description{
Binary raster to identify the study area (1 or NA)
}
\keyword{datasets}
