% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfire.rege.r, R/wildfires.r
\name{postfire.rege}
\alias{postfire.rege}
\alias{wildfires}
\title{Simulation of post-fire regeneration in forest areas}
\usage{
postfire.rege(land, params)

wildfires(land, params, out.maps = FALSE, verbose = FALSE)
}
\arguments{
\item{land}{Data frame with the description of the landscape: cell identifier, land-cover type, and
time since last cover-change}

\item{params}{List of the parameters of the model}

\item{out.maps}{A flat to save maps as .tif files of wildfire identifier, wildfire step and fire risk}

\item{orography}{Data frame with the geographic data of the raster cells: cell identifier, region 
(\code{GZ} is Galicia and \code{PT} is Protugal), protected (\code{TRUE}/\code{FALSE}), elevation (in m), aspect
 (1 - north, 2 - east, 3 - south, 4 - west), dens.road (density of roads per km^2^), slope (in %), and utm 
 (code of the UTM 1x1 grid).}
}
\value{
Returns a vector with cells' identifiers to change the land-cover type

Returns a vector with cells' identifiers burnt, suppresed
}
\description{
Simulation of post-fire regeneration in forest areas

Simulation of wildfires
}
\examples{
data(landscape)
data(orography)
data(postfire.trans.matrix)
params = default.params()
landscape$trans.type = NA 
x = landscape$cell.id[landscape$lct \%in\% c("pine", "oak")]
y = sample(x, 100, replace=F)
landscape$trans.type[landscape$cell.id \%in\% y] = "Fire"
landscape$tschg[landscape$cell.id \%in\% y] = 0
res = postfire.rege(landscape, orography, params)

data(landscape)
params = default.params()
res = wildfires(landscape, params, out.maps=F, verbose=F)

}
