% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{postfire.trans.matrix}
\alias{postfire.trans.matrix}
\title{Fire spread type and Main wind direction}
\format{
A data frame with as many rows as grid cells in the study area and 6 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{pfst.wind}{Fire spread type is wind-driven}
  \item{pfst.topo}{Fire spread type is topographic}
  \item{pwind.n}{Probability (0-100) of north winds dominance}
  \item{pwind.nw}{Probability (0-100) of north-west winds dominance}
  \item{pwind.w}{Probability (0-100) of west winds dominance}
}
}
\usage{
postfire.trans.matrix
}
\description{
FST & Probability wind directions
}
\keyword{datasets}
