% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescribed.burn.r
\name{prescribed.burn}
\alias{prescribed.burn}
\title{Simulation of prescribed burns}
\usage{
prescribed.burn(land, params, out.maps = FALSE, verbose = FALSE)
}
\arguments{
\item{land}{Data frame with the description of the landscape: cell identifier, land-cover type, and
time since last cover-change}

\item{params}{List of the parameters of the model}

\item{out.maps}{A flat to save maps as .tif files of wildfire identifier, wildfire step and fire risk}
}
\value{
Returns a vector with cells' identifiers burnt, suppressed
}
\description{
Simulation of prescribed burns
}
\examples{
data(landscape)
params = default.params()
res = wildfires(landscape, params, out.maps=F, verbose=F)

}
