% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMAINS.r
\docType{package}
\name{REMAINS}
\alias{REMAINS}
\title{REMAINS: A package for simulating landscape dynamics in mountain landscapes}
\description{
The REMAINS package provides a set of functions to simulate the main landscape-level processes 
driving fire-vegetation dynamics in mountain landscapes of northern Iberian Peninsula.
It includes the main anthropogenic and natural (abiotic and biotic) drivers of landscape change to study 
their spatio-temporal interactions and feedback effects.
}
\section{Author(s)}{

\bold{Maintainer}: Núria Aquilué \email{nuria.aquilue@ctfc.cat}  

\bold{Authors}: Núria Aquilué, Silvana Pais, Adrián Regos
}

