% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{burnt.area.dist.gz}
\alias{burnt.area.dist.gz}
\title{Timber harvesting allocation}
\format{
A data frame with as many rows as grid cells in the study area and 6 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{enpe}{Code of the protection status: 0 - none, 1 - national park, 2 - natural park, 
  3 - Natural place of national interest, 4 - Wildlife nature reserve, 5 - Integral natural reserve,
  6 - Partial natural reserve, 7 - Peripheral protection zone of a national park, 
  8 - Peripheral protection zone of a natural park, 9 - Peripheral protection zone of a natural place}
  \item{dist.path}{Distance to the nearest road or forest track, in m}
  \item{slope.pctg}{Stand slope, in /100}
  \item{dist.industry}{Distance to the nearest wood transformation industry, in km}
  \item{dist.biomass}{Distance to the nearest biomass plant, in km}
}
}
\usage{
burnt.area.dist.gz
}
\description{
Landscape variables influencing the allocation of sylvicultural prescriptions
}
\keyword{datasets}
