% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.risk.r
\name{fire.risk}
\alias{fire.risk}
\title{Compute fire risk as a function of landscape susceptibility, danger, and damage}
\usage{
fire.risk(land, ha.cell)
}
\arguments{
\item{land}{Data frame with the description of the landscape: cell identifier, land-cover type, and
time since last cover-change}

\item{ha.cell}{Number of hectares each raster cell in the grid represents}
}
\value{
Returns a vector with the fire risk of the raster cells that is function fo the land-cover type and 
the orography of the landscape
}
\description{
Compute fire risk as a function of landscape susceptibility, danger, and damage
}
\examples{
data(landscape)
res = fire.risk(landscape, 1)

}
