% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fire.size.dist.gz}
\alias{fire.size.dist.gz}
\title{Fire size distribution in Galicia}
\format{
A data frame with two variables:
\describe{
  \item{upper.th}{Upper threshold of discrete intervals of fire size}
  \item{freq}{Number of fires per class}
}
}
\usage{
fire.size.dist.gz
}
\description{
Discrete distribution of annual area burnt in Galicia observed between 1986-2011
}
\keyword{datasets}
