% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.cover.change.r
\name{land.cover.change}
\alias{land.cover.change}
\title{Simulation of land-cover changes}
\usage{
land.cover.change(
  land,
  params,
  lc.trans = 1,
  trgt.dmnd = 0,
  visit.cells = NULL
)
}
\arguments{
\item{land}{Data frame with the description of the landscape: cell identifier, land-cover type, and
time since last cover-change}

\item{params}{List of the parameters of the model}

\item{lc.trans}{Identifier of the land-cover transition: 1 - smart plantation, 2 - agriculture conversion, 
3 - rural abandonment, and 4 - pasture abandonment.}

\item{trgt.dmnd}{Number of hectares to be converted to the target land-cover type}

\item{visit.cells}{Vector of cells' identifier that have already been subject of change for a land-cover change process}
}
\value{
Returns a vector with cells' identifiers to change the land-cover type
}
\description{
Simulation of land-cover changes
}
\examples{
data(landscape)
params = default.params()
res = land.cover.change(landscape, params, 1, 100, NULL)

}
