% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.dyn.mdl.r
\name{land.dyn.mdl}
\alias{land.dyn.mdl}
\title{REMAINS model}
\usage{
land.dyn.mdl(
  scenDir,
  is.land.cover.change = TRUE,
  is.wildfire = TRUE,
  is.prescribed.burn = TRUE,
  is.postfire.rege = TRUE,
  is.forest.recover = TRUE,
  is.afforestation = TRUE,
  is.encroachment = TRUE,
  nrun = 1,
  params = NULL,
  lcc.demand = NULL,
  save.land = FALSE,
  out.maps = FALSE,
  write.outputs = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{scenDir}{String with the name of the scenario that is the directory path to save the model's outputs}

\item{is.land.cover.change}{A flag to indicate that land cover changes are simulated}

\item{is.wildfire}{A flag to indicate that wildfires are simulated}

\item{is.prescribed.burn}{A flag to indicate that prescribed burns are simulated}

\item{is.postfire.rege}{A flag to indicate that post-fire regeneration is simulated}

\item{is.forest.recover}{A flag to indicate that forest recover after fire is simulated}

\item{is.afforestation}{A flag to indicate that afforestation is simulated}

\item{is.encroachment}{A flag to indicate that encroachment is simulated}

\item{nrun}{Number of replicates to run the model}

\item{params}{A list of the default model's parameters}

\item{lcc.demand}{A data frame with the area to be changed by each land-cover transition per time step}

\item{save.land}{A flag to save as a RDS file the \code{landscape} data frame at the time step indicated in \code{out.seq}}

\item{out.maps}{A flat to save maps as .tif files of wildfire identifier, wildfire step and fire risk}

\item{write.outputs}{A flag to write the results of the landscape processes as text files}

\item{verbose}{A flag to prompt extra information}
}
\value{
A list with the following 18 items:
 \itemize{
   \item{\code{Land}: A data frame with the land-cover types areas, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{lct}: Land-cover type.}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{lcc}: A data frame of land-cover transformations or conversions 
   (included if \code{is.land.cover.change}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{trans.type}: Transformation type: Rural abandonment, Pasture abandonment, Agricultural conversion, Pasture conversion and/or Smart plantation.}
        \item{\code{region}: Region: Portugal - PT and Spain - GZ.}
        \item{\code{area}: Area (in ha).}
        \item{\code{trgt.dmnd}: Target demand (in ha).}
      }
   }
   \item{\code{Fires}: A data frame of target, burnt and suppressed area per wildfire
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{region}: Event locations: Portugal and Spain.}
        \item{\code{fire.id}: Fire event identificator.}
        \item{\code{atarget}: Target area to be burnt (in ha).}
        \item{\code{aburnt}: Area burnt (in ha).}
        \item{\code{asupp.fuel}: Area suppressed in low fire spread conditions (in ha).} 
        \item{\code{asupp.mosaic}: Area suppressed due to the suppression opportunities created by croplands (in ha).} 
        \item{\code{rem}: Remaining area, not burnt, neither suppressed (in ha).}    
      }
   }
   \item{\code{lc.burnt.supp}: A data frame of burned and suppressed land-cover types 
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{region}: Region: Portugal and Spain.}
        \item{\code{lct}: Land-cover type.}
        \item{\code{target}: Target area to be burnt (in ha).} 
        \item{\code{burnt}: Area effectively burnt (in ha).}
        \item{\code{supp}: Area effectively suppressed (in ha).}
      }
   }   
    \item{\code{lc.burnt}: A data frame of burned land-cover types 
   (included if \code{is.wildfires}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{region}: Region: Portugal and Spain.}
        \item{\code{lct}: Land-cover type.}
        \item{\code{target}: Target area to be burnt (in ha).} 
        \item{\code{burnt}: Area effectively burnt (in ha).}
        \item{\code{supp}: Area effectively suppressed (in ha).}
      }
   }
   \item{\code{unburnt.land}: A data frame of unburned area by land-cover type
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{lct}: Land-cover type.}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{Forest.age.burnt}: A data frame with the count burned pines and oaks by region and age
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{region}: Region: Portugal and Spain.}
        \item{\code{lct}: Land-cover type.}
        \item{\code{age}: Age (in years).}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{Forest.age.supp}: A data frame with the count suppressed pines and oaks by region and age
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{region}: Region: Portugal and Spain.}
        \item{\code{lct}: Land-cover type.}
        \item{\code{age}: TAge (in years).}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{unburnt.aok.age}: A data frame with unburnt oak forest per age
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{age}: Age (in years).}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{unburnt.pine.age}: A data frame with unburnt pine forest per age 
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{age}: Age (in years).}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{pb}: A data frame with the area burned by prescribed fire
   (included if \code{is.prescribed.burn}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{fire.id}: Fire event identificator.}
        \item{\code{atarget}: Target area to be burnt (in ha).}
        \item{\code{aburnt}: Area effectively burnt by prescribed fires(in ha).}
      }
   }
   \item{\code{lc.pburnt}: A data frame with prescribed fire areas by land-cover types 
   (included if \code{is.prescribed.burn}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{lct}: Land-cover type burned by prescribed fire.}
        \item{\code{tschg}: Time since the last land-cover transition (in years).}
        \item{\code{aburnt}: Area effectively burnt (in ha).}
      }
   }
   \item{\code{postfire.rege}: A data frame of post-fire regeneration by land-cover type
   (included if \code{is.postfire.rege}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{lct}: Land-cover type.}
        \item{\code{post.lct}: Land-cover type post-fire.}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{Forest.recover}: A data frame of potential forest recovering from shrublands
   (included if \code{is.forest.recover}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{current.shrub.to.pine}: Burnt area of shrubland in the current time step that will convert to pine forest in the following years (in ha).}
        \item{\code{current.shrub.to.oak}: Burnt area of shrubland in the current time step that will convert to oak forest in the following years (in ha).}
        \item{\code{convert.to.pine}: Shrubland area that convert/transform to pine forest (in ha).}
        \item{\code{convert.to.oak}: Shrubland area that convert/transform to oak forest (in ha).}
      }
   }
    \item{\code{Afforest}: A data frame with the colonization of shrublands by oak and/or pines species
   (included if \code{is.afforestation}), with columns:
     \itemize{
        \item{\code{year}: Year YYYY.}
        \item{\code{potential.shrub}: Area (in ha).}
        \item{\code{target.afforest.oak}: Area (in ha).}
        \item{\code{target.afforest.pine}: Area (in ha).}
        \item{\code{afforest.oak}: Area (in ha).}
        \item{\code{afforest.oak}: Area (in ha).}
      }
   }
   \item{\code{oak.age}: A data frame of oak forest per age with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{age}: Age (in years).}
        \item{\code{area}: Area (in ha).}
      }
   }
   \item{\code{pine.age}: A data frame of pine forest per age with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{age}: Age (in years).}
        \item{\code{area}: Area (in ha).}
      }
   }
    \item{\code{Encroach}: A data frame of new shrubland area following encroachment
   (included if \code{is.encroachment}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{potential.rocky}: Area (in ha).}
        \item{\code{target.encroach}: Area (in ha).}
        \item{\code{encroach}: Area (in ha).}
      }
    }
 }
}
\description{
Run the REMAINS model that includes the processes of land-cover changes, wildfires, prescribed fires, 
and vegetation dynamics in mountain landscapes.
}
\examples{

\dontrun{
library(REMAINS) 
# Run one single 2020 replicate with forest management 
result = land.dyn.mdl(is.harvest = T)
}

}
