% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{landscape}
\alias{landscape}
\title{Forest landscape features of the Biosphere Reserve Ger?s-Xur?s}
\format{
A data frame with as many rows as grid cells in the study area and 3 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{lct}{Tree species or land-cover types: pine, oak, water, shrub, sparseveg, crop, grass, urban}
  \item{tschg}{Time since change (in years)}
}
}
\usage{
landscape
}
\description{
Landscape and forest stands characteristics of the study area in 2020 at 1 ha of spatial resolution
}
\keyword{datasets}
