% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lcc.demand}
\alias{lcc.demand}
\title{Mask of the study area}
\format{
Raster of 2800 (nrow) x 2009 (ncol)
\describe{
Raster of the study area (1 or NA) in the UTM31N - ETRS89 projection, at 1 ha of spatial resoltuion. 
The unique grid cell identificator \code{cell.id} in the \code{landscape} data frame coincides with 
the position of the location in the \code{mask} raster.
}
}
\usage{
lcc.demand
}
\description{
Binary raster to identify the study area (1 or NA)
}
\keyword{datasets}
