% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{postfire.trans.matrix}
\alias{postfire.trans.matrix}
\title{Post-fire transitions matrix}
\format{
A data frame with as many rows and columns as land-cover types in the study area:
\describe{
  \item{lc.prefire}{Land-cover types}
  \item{crop}{Croplands}
  \item{grass}{Grasslands}
  \item{shrub}{Shrublands}
  \item{pine}{Pine forest}
  \item{oak}{Oak forest}
  \item{sparseveg}{Sparse vegetation and rocky areas}
  \item{water}{Water}
  \item{urban}{Urban}
  \item{baresoil}{baresoil}
}
}
\usage{
postfire.trans.matrix
}
\description{
Post-fire transition rates calibrated for the study area
}
\keyword{datasets}
